guards applied to users and workspace api left for testing

+++ src/Users/users.service.ts
@@ -1,4 +1,4 @@
-import { ConflictException, HttpException, Injectable, InternalServerErrorException, NotFoundException } from "@nestjs/common";
+import { ConflictException, Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { UsersSchema } from "./users.schema";
 import mongoose, { Model } from "mongoose";
@@ -10,8 +10,10 @@ export class UsersService {
     constructor(@InjectModel(UsersSchema.name) private usersModel: Model<UsersSchema>) { }
 
     //Add users-----------------------------------------------------------------------
-    async addUser({ password, ...userData }: UsersDto) {
+    async addUser({ password, ...userData }: UsersDto, req: any) {
         try {
+            const isAdmin = req.users.isAdmin;
+            if (!isAdmin) throw new UnauthorizedException;
             const foundUser = await this.usersModel.findOne({ $and: [{ email: userData.email }, { workspaceId: userData.workspaceId }] }).exec();
             if (foundUser) throw new ConflictException; //409
             bcrypt.hash(password, 10, async (e, hashedPass) => {
@@ -23,7 +25,7 @@ export class UsersService {
                     const newUser = new this.usersModel({ password: hashedPass, ...userData });
                     const savedUser = await newUser.save();
                     const { isAdmin, ...savedUserDto } = savedUser;
-                    return { ...savedUserDto };
+                    return { savedUserDto };
                 }
             })
         }
@@ -34,8 +36,10 @@ export class UsersService {
     }
 
     //Get all users------------------------------------------------------------------
-    async getAllUsers() {
+    async getAllUsers(req: any) {
         try {
+            const isAdmin = req.users.isAdmin;
+            if (!isAdmin) throw new UnauthorizedException;
             const allUsers = await this.usersModel.find({}, { _id: 1, email: 1, role: 1, workspaceId: 1 }).exec();
             return allUsers;
         }
@@ -46,8 +50,10 @@ export class UsersService {
     }
 
     //Get users by id----------------------------------------------------------------
-    async getUserById(userId: mongoose.Schema.Types.ObjectId) {
+    async getUserById(userId: mongoose.Schema.Types.ObjectId, req: any) {
         try {
+            const isAdmin = req.users.isAdmin;
+            if (!isAdmin) throw new UnauthorizedException;
             const findUser = await this.usersModel.findOne({ _id: userId }, { _id: 1, email: 1, role: 1, workspaceId: 1 }).exec();
             if (!findUser) throw new NotFoundException;
             return findUser;
@@ -59,8 +65,10 @@ export class UsersService {
     }
 
     //delete a user-----------------------------------------------------------------
-    async deleteUser(userId: mongoose.Schema.Types.ObjectId) {
+    async deleteUser(userId: mongoose.Schema.Types.ObjectId, req: any) {
         try {
+            const isAdmin = req.users.isAdmin;
+            if (!isAdmin) throw new UnauthorizedException;
             const deletedUser = await this.usersModel.findOneAndDelete({ _id: userId }, { returnDocument: "after" }).exec()
             if (!deletedUser) throw new NotFoundException;
             return { deleted_userId: deletedUser._id, deleted_email: deletedUser.email, deleted_role: deletedUser.role, deleted_workspaceId: deletedUser.workspaceId };
@@ -72,8 +80,10 @@ export class UsersService {
     }
 
     //edit a user---------------------------------------------------------------------
-    async editUser(userId: mongoose.Schema.Types.ObjectId, updateUserDto: updateUserDto) {
+    async editUser(userId: mongoose.Schema.Types.ObjectId, updateUserDto: updateUserDto, req: any) {
         try {
+            const isAdmin = req.users.isAdmin;
+            if (!isAdmin) throw new UnauthorizedException;
             const editedUser = await this.usersModel.findOneAndUpdate({ _id: userId }, { updateUserDto }, { returnDocument: "after" }).exec();
             if (!editedUser) throw new NotFoundException;
             return { edit_userId: editedUser._id, edit_email: editedUser.email, edit_role: editedUser.role, edit_workspaceId: editedUser.workspaceId };
